refactor: consolidate resolver micro-services 51 -> 35 files
Move satellite types into their natural host files as additional
top-level package-private classes — valid Java, zero caller changes.

Merges:
  FlowState         -> FlowRecord.java
  PendingDeliveryState -> PendingDeliveryTracker.java
  WorkerAvailabilityGate -> PendingStateDecisionService.java
  ResolverPendingState, PostCreationUpdateService -> PendingDeliveryCreationService.java
  StackMetrics, OutstandingNeededService,
    StockResolver, WarehouseCountService -> StockSnapshot.java
  FlowTimeoutCleanupService, TickPendingTelemetryService -> TickPendingService.java
  ReservationSyncService -> ReservationReleaseService.java
  DeliveryChildRecoveryService -> DeliveryChildLifecycleService.java
  ResolverRecheck    -> ResolverCooldown.java
  DeliveryCallbackService -> RequestResolver.java
  DeliveryOriginMatcher -> ChildReconciliationService.java

No logic changes.
