--- src/main/java/com/thesettler_x_create/minecolonies/requestsystem/resolver/CreateShopDeliveryCallbackService.java
@@ -1,33 +1,0 @@
-package com.thesettler_x_create.minecolonies.requestsystem.resolver;
-
-import com.minecolonies.api.colony.requestsystem.manager.IRequestManager;
-import com.minecolonies.api.colony.requestsystem.request.IRequest;
-
-/** Routes MineColonies delivery callbacks to the matching local Create Shop resolver instance. */
-final class CreateShopDeliveryCallbackService {
-  void onDeliveryCancelled(IRequestManager manager, IRequest<?> request) {
-    CreateShopRequestResolver resolver =
-        CreateShopDeliveryResolverLocator.findResolverForDelivery(manager, request);
-    if (resolver == null) {
-      resolver = CreateShopDeliveryResolverLocator.findResolverByDeliveryToken(manager, request);
-    }
-    if (resolver != null) {
-      resolver.handleDeliveryCancelled(manager, request);
-      return;
-    }
-    CreateShopDeliveryResolverLocator.logUnresolvedDeliveryCallback("cancelled", manager, request);
-  }
-
-  void onDeliveryComplete(IRequestManager manager, IRequest<?> request) {
-    CreateShopRequestResolver resolver =
-        CreateShopDeliveryResolverLocator.findResolverForDelivery(manager, request);
-    if (resolver == null) {
-      resolver = CreateShopDeliveryResolverLocator.findResolverByDeliveryToken(manager, request);
-    }
-    if (resolver != null) {
-      resolver.handleDeliveryComplete(manager, request);
-      return;
-    }
-    CreateShopDeliveryResolverLocator.logUnresolvedDeliveryCallback("complete", manager, request);
-  }
-}
