--- src/main/java/com/thesettler_x_create/minecolonies/requestsystem/resolver/CreateShopOutstandingNeededService.java
@@ -1,19 +1,0 @@
-package com.thesettler_x_create.minecolonies.requestsystem.resolver;
-
-import com.minecolonies.api.colony.requestsystem.request.IRequest;
-import com.minecolonies.api.colony.requestsystem.requestable.IDeliverable;
-import com.minecolonies.api.colony.requestsystem.requestable.INonExhaustiveDeliverable;
-
-/** Computes outstanding required amount for a request after reservation/non-exhaustive offsets. */
-final class CreateShopOutstandingNeededService {
-  int compute(IRequest<?> request, IDeliverable deliverable, int reservedForRequest) {
-    if (request == null || deliverable == null) {
-      return 0;
-    }
-    int needed = deliverable.getCount();
-    if (deliverable instanceof INonExhaustiveDeliverable nonExhaustive) {
-      needed -= nonExhaustive.getLeftOver();
-    }
-    return Math.max(0, needed - Math.max(0, reservedForRequest));
-  }
-}
