+++ src/main/java/com/thesettler_x_create/minecolonies/requestsystem/resolver/CreateShopPendingStateDecisionService.java
@@ -127,3 +127,18 @@ final class CreateShopPendingStateDecisionService {
     }
   }
 }
+
+/** Worker-availability decisions for request ordering and pending resume. */
+final class CreateShopWorkerAvailabilityGate {
+  boolean shouldDeferNetworkOrder(boolean workerWorking, int neededCount) {
+    return !workerWorking && neededCount > 0;
+  }
+
+  boolean shouldResumePending(boolean workerWorking, int pendingCount) {
+    return pendingCount > 0;
+  }
+
+  boolean shouldKeepPendingState(boolean workerWorking, int pendingCount) {
+    return !workerWorking && pendingCount > 0;
+  }
+}
