+++ src/main/java/com/thesettler_x_create/minecolonies/requestsystem/resolver/CreateShopTickPendingService.java
@@ -1,6 +1,9 @@
 package com.thesettler_x_create.minecolonies.requestsystem.resolver;
 
+import com.minecolonies.api.colony.requestsystem.management.IRequestHandler;
 import com.minecolonies.api.colony.requestsystem.manager.IRequestManager;
+import com.minecolonies.api.colony.requestsystem.request.IRequest;
+import com.minecolonies.api.colony.requestsystem.requestable.IDeliverable;
 import com.minecolonies.api.colony.requestsystem.token.IToken;
 import com.minecolonies.core.colony.requestsystem.management.IStandardRequestManager;
 import com.thesettler_x_create.Config;
@@ -130,5 +133,122 @@ final class CreateShopTickPendingService {
     }
     flowTimeoutCleanupService.processTimedOutFlows(resolver, standardManager, level);
     tickPendingTelemetryService.recordAndMaybeLogPerf(level, System.nanoTime() - perfStart);
+  }
+}
+
+/** Handles timeout-driven resolver cleanup for stale non-terminal flow records. */
+final class CreateShopFlowTimeoutCleanupService {
+  private final CreateShopRequestStateMutatorService requestStateMutatorService;
+
+  CreateShopFlowTimeoutCleanupService(
+      CreateShopRequestStateMutatorService requestStateMutatorService) {
+    this.requestStateMutatorService = requestStateMutatorService;
+  }
+
+  void processTimedOutFlows(
+      CreateShopRequestResolver resolver, IStandardRequestManager manager, Level level) {
+    if (resolver == null || manager == null || level == null) {
+      return;
+    }
+    long timeout = resolver.getInflightTimeoutTicksSafe();
+    for (CreateShopFlowRecord record :
+        resolver.getFlowStateMachine().collectTimedOut(level.getGameTime(), timeout)) {
+      IToken<?> token = record.getRequestToken();
+      boolean runtimeDeliveryWindowOpen =
+          resolver.getParentDeliveryTokensSnapshot().contains(token)
+              || resolver.hasDeliveriesCreated(token)
+              || resolver.getPendingTracker().hasDeliveryStarted(token);
+      IRequest<?> request = null;
+      try {
+        request = manager.getRequestHandler().getRequest(token);
+      } catch (Exception ignored) {
+        // Missing requests are cleaned up below.
+      }
+      if (request == null && runtimeDeliveryWindowOpen) {
+        // Request graph lookups can be transiently stale; do not clear active parent lifecycle.
+        resolver.touchFlow(token, level.getGameTime(), "timeout-cleanup:skip-runtime-active");
+        continue;
+      }
+      if (request != null) {
+        boolean terminal = CreateShopRequestResolver.isTerminalRequestState(request.getState());
+        boolean deliveryWindowOpen = request.hasChildren() || runtimeDeliveryWindowOpen;
+        if (!terminal && deliveryWindowOpen) {
+          // Active deliveries can outlive local flow timestamps; do not clear parent lifecycle
+          // here.
+          resolver.touchFlow(token, level.getGameTime(), "timeout-cleanup:skip-active-delivery");
+          continue;
+        }
+        resolver.transitionFlow(
+            manager,
+            request,
+            CreateShopFlowState.FAILED,
+            "timeout-cleanup",
+            record.getStackLabel(),
+            record.getAmount(),
+            "com.thesettler_x_create.message.createshop.flow_timeout");
+        if (request.getRequest() instanceof IDeliverable) {
+          resolver.releaseReservation(manager, request);
+        }
+      }
+      requestStateMutatorService.clearPendingTokenState(resolver, manager, token, true);
+    }
+  }
+}
+
+/** Rate-limited tick-pending diagnostics and perf logging state. */
+final class CreateShopTickPendingTelemetryService {
+  private final Object debugLock = new Object();
+  private volatile long lastTickPendingDebugTime = 0L;
+  private long lastPerfLogTime = 0L;
+  private long lastTickPendingNanos = 0L;
+
+  boolean shouldLogTickPending(Level level) {
+    long now = level.getGameTime();
+    if (now == 0L) {
+      return true;
+    }
+    synchronized (debugLock) {
+      if (now - lastTickPendingDebugTime >= Config.TICK_PENDING_DEBUG_COOLDOWN.getAsLong()) {
+        lastTickPendingDebugTime = now;
+        return true;
+      }
+      return false;
+    }
+  }
+
+  void logTickPendingCandidates(IRequestHandler requestHandler, Set<IToken<?>> pendingTokens) {
+    int logged = 0;
+    for (IToken<?> token : List.copyOf(pendingTokens)) {
+      if (logged >= 5) {
+        break;
+      }
+      try {
+        IRequest<?> req = requestHandler.getRequest(token);
+        String type = req == null ? "<null>" : req.getRequest().getClass().getName();
+        String state = req == null ? "<null>" : req.getState().toString();
+        TheSettlerXCreate.LOGGER.info(
+            "[CreateShop] tickPending: candidate {} type={} state={}", token, type, state);
+        logged++;
+      } catch (IllegalArgumentException ignored) {
+        // Missing request.
+      }
+    }
+  }
+
+  void recordAndMaybeLogPerf(Level level, long tickPendingNanos) {
+    lastTickPendingNanos = tickPendingNanos;
+    if (!Config.DEBUG_LOGGING.getAsBoolean()) {
+      return;
+    }
+    if (level == null) {
+      return;
+    }
+    long now = level.getGameTime();
+    if (now != 0L && now - lastPerfLogTime < Config.PERF_LOG_COOLDOWN.getAsLong()) {
+      return;
+    }
+    lastPerfLogTime = now;
+    TheSettlerXCreate.LOGGER.info(
+        "[CreateShop] perf tickPending={}us", lastTickPendingNanos / 1000L);
   }
 }
